requirements.txt
matplotlib==3.4.2
Shapely==1.7.1
validators==0.18.2
omegaconf==2.1.0rc1
hydra_core==1.1.0rc1
opencv_python==4.5.2.52
torchvision==0.8.2
fvcore==0.1.5.post20210518
streamlit==0.82.0
av==8.0.3
pafy==0.5.5
termcolor==1.1.0
cityscapesScripts==2.2.0
psutil==5.7.2
starlette==0.14.2
wandb==0.10.31
recommonmark==0.7.1
sphinx_rtd_theme==0.5.2
Flask==1.1.2
Pygments==2.6.1
Sphinx==2.4.4
torch==1.7.1
iopath==0.1.8
docutils==0.16
numpy==1.19.5
seaborn==0.11.1
tqdm==4.61.0
dataclasses==0.6
scipy==1.6.3
tabulate==0.8.9
coremltools==4.1
mmdet==2.12.0
black==21.4b2
mmcv==1.3.5
cloudpickle==1.6.0
requests==2.25.1
pydantic==1.8.2
onnx==1.9.0
pandas==1.2.4
pycocotools==2.0.2
caffe2==0.8.1
fastapi==0.65.1
hydra==2.5
Pillow==8.2.0
protobuf==3.17.2
PyYAML==5.4.1
thop==0.0.31-2005241907
